Added functionality for increasing and decreasing the quantity of goods in the basket, improved card UI, fixed bug with data updationg, added unfinished footer

+++ server/services/basketService.js
@@ -2,16 +2,42 @@ const ApiError = require("../error/ApiError");
 const { Basket, Product, User } = require("../models/models");
 
 class BasketService {
-  async addToBasket(userId, productId) {
+  async addToBasket(userId, productId, amount = 1) {
     if (!productId) {
       throw ApiError.badRequest("Product ID is required");
     }
-    const basket = await Basket.create({ userId, productId });
+
+    let basket = await Basket.findOne({ where: { userId, productId } });
+    if (basket) {
+      basket.quantity += amount;
+      await basket.save();
+    } else {
+      basket = await Basket.create({ userId, productId, quantity: amount });
+    }
+
     return basket;
   }
 
   async removeFromBasket(userId, productId) {
-    const basket = await Basket.destroy({ where: { userId, productId } });
+    const basket = await Basket.findOne({ where: { userId, productId } });
+    if (!basket) {
+      throw ApiError.badRequest("Product is not in basket");
+    }
+
+    await Basket.destroy({ where: { userId, productId } });
+    return basket;
+  }
+
+  async updateBasketItem(userId, productId, newQuantity) {
+    if (newQuantity < 1) {
+      throw ApiError.badRequest("Quantity must be at least 1");
+    }
+    const basket = await Basket.findOne({ where: { userId, productId } });
+    if (!basket) {
+      throw ApiError.badRequest("Product is not in basket");
+    }
+    basket.quantity = newQuantity;
+    await basket.save();
     return basket;
   }
 
@@ -21,7 +47,7 @@ class BasketService {
         include: {
           model: Product,
           as: "basketProducts",
-          through: { attributes: [] },
+          through: { attributes: ["quantity"] },
         },
       });
